Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.8
fastapi-cli==0.0.21
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastjsonschema==2.21.2
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
nbclient==0.10.2
nbconvert==7.17.0
nbformat==5.10.4
openai==2.21.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
rpds-py==0.27.1
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
motor=3.7.1
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.49.3
tinycss2==1.4.0
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.1.1
yarg==0.1.9
zipp==3.23.0
